+++ src/client/components/ui/ModalHeader.ts
@@ -56,6 +56,7 @@ export const modalHeader = ({
           @click=${onBack}
           class="${buttonClass}"
           aria-label="${ariaLabel}"
+          data-modal-back
         >
           <svg
             xmlns="http://www.w3.org/2000/svg"